Initial support for annotating multiple frames
The host page annotator is split into a host and guest version. The
`Annotator.Guest` class implements most of the behavior from the
old `Annotator.Host`, but leaves out everything to do with the widget
hosting aspects (opening, closing and coordinating drag resize with
the sidebar frame).

The bridge is improved to support multiple connections and multiplex
RPC calls. Connections are established by probing other frames to
discovered other bridge annotators. A frame may act like a DHCP
server, responding to discovery probes and offering to connect.

The annotator service is refactored to behave more reasonably with
service and dynamic frame discovery. Rather than assuming that
document metadata is fetched from a single frame, entities from
all frames are registered when they connect. The Store plugin is
patched to load only links that have not previously been loaded
so that API calls can be made as additional entities are registered.

Most functionality, aside from highlight mode, is broken. This is a
result of the difficulty of dealing with the heatmap when it may be
hard or impossible to determine the real offsets of the annotated
frames relative to the offset of a global heatmap. The changes to
extract the heatmap were left out of this commit in order to keep
the size of an already large commit down.